\name{concentration_mg_ml}
\alias{concentration_mg_ml}
\title{Calculate solution concentration}
\description{Calculates concentration from peptide mass and final solution volume.}
\usage{concentration_mg_ml(mass_mg, volume_ml)}
\arguments{
  \item{mass_mg}{Peptide mass in milligrams.}
  \item{volume_ml}{Final solution volume in millilitres.}
}
\value{Concentration in milligrams per millilitre.}
\examples{
concentration_mg_ml(10, 5)
}
